refactor(desktop): extract session tool-part module and harden event coverage
